Refactor authentication handling and configuration; add logging for server start
