+++ configs/config.go
@@ -8,33 +8,26 @@ import (
 )
 
 type Config struct {
-	Db   DbConfig
 	Auth AuthConfig
 }
 
-type DbConfig struct {
-	Dsn string
-}
-
 type AuthConfig struct {
-	ClientSecret string
 	ClientId     string
+	ClientSecret string
+	ClientUrl    string
 }
 
 func LoadConfig() *Config {
 	err := godotenv.Load()
-
 	if err != nil {
 		log.Println("Error loading .env file")
 	}
 
 	return &Config{
-		Db: DbConfig{
-			Dsn: os.Getenv("DSN"),
-		},
 		Auth: AuthConfig{
-			ClientSecret: os.Getenv("CLIENT_SECRET"),
 			ClientId:     os.Getenv("CLIENT_ID"),
+			ClientSecret: os.Getenv("CLIENT_SECRET"),
+			ClientUrl:    os.Getenv("CLIENT_URL"),
 		},
 	}
 }
