+++ pkg/codec/decode.go
@@ -1,4 +1,4 @@
-package req
+package codec
 
 import (
 	"encoding/json"
@@ -8,7 +8,6 @@ import (
 func Decode[T any](body io.ReadCloser) (T, error) {
 	var payload T
 	err := json.NewDecoder(body).Decode(&payload)
-
 	if err != nil {
 		return payload, err
 	}
